refactor it test, fix ce
