--- src/it/scala/com/wavesplatform/it/IntegrationTestsScheme.scala
@@ -1,9 +1,0 @@
-package com.wavesplatform.it
-
-import vsys.account.AddressScheme
-
-trait IntegrationTestsScheme {
-  AddressScheme.current.value = new AddressScheme {
-    override val chainId: Byte = 'I'
-  }
-}
