--- src/it/scala/com/wavesplatform/it/network/client/package.scala
@@ -1,7 +1,0 @@
-package com.wavesplatform.it.network
-
-import io.netty.channel.Channel
-
-package object client {
-  def id(chan: Channel): String = s"[${chan.id().asShortText()}]"
-}
