+++ src/it/scala/vsys/it/IntegrationTestsScheme.scala
@@ -1,0 +1,9 @@
+package vsys.it
+
+import vsys.account.AddressScheme
+
+trait IntegrationTestsScheme {
+  AddressScheme.current.value = new AddressScheme {
+    override val chainId: Byte = 'I'
+  }
+}
