+++ src/it/scala/vsys/it/network/client/package.scala
@@ -1,0 +1,7 @@
+package vsys.it.network
+
+import io.netty.channel.Channel
+
+package object client {
+  def id(chan: Channel): String = s"[${chan.id().asShortText()}]"
+}
