+++ src/it/scala/vsys/it/util/package.scala
@@ -1,0 +1,28 @@
+package vsys.it
+
+import io.netty.util.Timer
+import vsys.settings.Constants
+
+import scala.concurrent.{ExecutionContext, Future, Promise}
+import scala.concurrent.duration.{FiniteDuration, MILLISECONDS}
+
+package object util {
+  implicit class TimerExt(val timer: Timer) extends AnyVal {
+    def schedule[A](f: => Future[A], delay: FiniteDuration): Future[A] = {
+      val p = Promise[A]
+      try {
+        timer.newTimeout(_ => p.completeWith(f), delay.toMillis, MILLISECONDS)
+      } catch {
+        case t: Throwable => p.failure(t)
+      }
+      p.future
+    }
+
+    def retryUntil[A](f: => Future[A], cond: A => Boolean, retryInterval: FiniteDuration)
+                     (implicit ec: ExecutionContext): Future[A] =
+      f.flatMap(v => if (cond(v)) Future.successful(v) else schedule(retryUntil(f, cond, retryInterval), retryInterval))
+  }
+  implicit class LongExt(val l: Long) extends AnyVal {
+    def vsys: Long = l * Constants.UnitsInVsys
+  }
+}
